Rework: Improve sqlite3 schema.
'Directory' has been renamed to 'dir_name' as it's probably more
representative of what it is. Also, the 'next_ep' and 'next_s' columns
have both been removed in favor of checking dynamically. It could be
added back in the future, but really, it doesn't make too much of a
difference performance wise to check once if a row exists.

Among the schema changes is the addition of the 'special' column. This
is determine via a regex which checks if the filename contains strings
such as: "OVA", "ED", "NCED", "OP", "NCOP", ... However, this addition
requires a rework in the data structure that the query returns, as it
can return the episode and season number, or the name of the special.